+++ src/glyphtune/output/__init__.py
@@ -1,5 +1,5 @@
 """Audio output."""
 
-from glyphtune.output.stream import MonoStream
+from glyphtune.output.stream import Stream
 
-__all__ = ["MonoStream"]
+__all__ = ["Stream"]