+++ CHANGELOG.md
@@ -9,10 +9,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- `glyphtune.arrays` module.
+- Output stream with an arbitrary number of channels (`glyphtune.output.Stream`).
+- Effects module (including base `Effect` object).
+- Stereo effects module (including `StereoPan`, `StereoLevels`, `StereoInterMix`, and `StereoDelay`).
+- `glyphtune.arrays` module (including `FloatArray` type alias and `is_stereo_signal` function).
 
 ### Changed
 
+- `Waveform.sample_arr` is now expected to be able to take as an argument an array of the shape `(channels, samples)` and return an array of the same shape.
+- `Waveform.sample_seconds` and `Waveform.sample_samples` now accept a `channels` option and can sample multi-channel audio, returning arrays of the shape `(channels, samples)`.
+- `Waveform.sample_seconds` and `Waveform.sample_samples` now sample stereo audio by default.
+- `PhaseModulation` no longer accepts the `frequency_modulation` option and is no longer able to perform frequency modulation.
 - `Square` objects no longer subclass `Pulse` and now subclass `PeriodicWave`.
 - Moved `FloatArray` from `glyphtune` module to `glyphtune.arrays`.
 - `PeriodicWave` now properly overrides `__eq__` and `__repr__`.
@@ -21,8 +28,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Removed
 
-- Calculus module (including DerivativeWaveform and IntegralWaveform)
-- Frequency modulation
+- `glyphtune.output.MonoStream` class (replaced by `glyphtune.output.Stream`).
+- Calculus module (`glyphtune.waveforms.calculus`, including `DerivativeWaveform` and `IntegralWaveform`).
+- Frequency modulation (`glyphtune.waveforms.modulation.frequency_modulate`).
 
 ### Fixed